// data/github/current/ui/packages/memex/src/client/api/memex/contracts.ts
export interface GetCustomTemplatesResponse {
  templates: Array<CustomTemplate>
  recommendedTemplates?: Array<CustomTemplate>
}

/** Represents the type for a GitHub system template */
export type SystemTemplate = {
  title: string
  id: string
  shortDescription?: string
  imageUrl: {
    light: string
    dark: string
  }
}